# chatbot-vf/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentación Diputación Alicante",
        "description": "Apis Alicante",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/getRoleData": {
            "get": {
                "tags": [
                    "roles"
                ],
                "summary": "Mostrar el listado de roles",
                "description": "Se muestra el listado de los registros solicitados.",
                "operationId": "be38a431fb1e49ea96bd6e93a4a2d9b3",
                "responses": {
                    "200": {
                        "description": "Mostrar todos los roles."
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
...
        "/api/getPermissionData": {
            "get": {
                "tags": [
                    "permissions"
                ],
                "summary": "Obtener todos los datos de permisos.",
                "description": "Devuelve todos los datos de permisos disponibles en el sistema.",
                "operationId": "6ce7e5fa6dcd9c55bd75407007656afa",
                "responses": {
                    "200": {
                        "description": "Datos de permisos obtenidos correctamente."
                    },
                    "422": {
                        "description": "Error al obtener los datos de permisos.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Error al obtener los datos de permisos"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/createRole": {
            "post": {
                "tags": [
                    "roles"
                ],
                "summary": "Crear un nuevo rol con los permisos proporcionados.",
                "description": "Crea un nuevo rol en la base de datos y asigna los permisos especificados.",
                "operationId": "ce8db63b7ace9d57d173d3ba4e98357f",
                "requestBody": {
                    "description": "Datos del rol a crear.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "permissions"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Rol Ejemplo"
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "El rol se creó exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rol creado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "El rol ya existe en la base de datos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "El rol ya existe en la base de datos"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/getRoleId/{id}": {
            "get": {
                "tags": [
                    "roles"
                ],
                "summary": "Obtiene un rol con sus permisos",
                "description": "Obtiene un rol con sus permisos.",
                "operationId": "e80126f6c6370298fa0bd3a0103cf929",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol con sus permisos obtenidos correctamente."
                    },
                    "422": {
                        "description": "Error al obtener el rol."
                    }
                }
            }
        },
        "/api/updateRole": {
            "post": {
                "tags": [
                    "roles"
                ],
                "summary": "Actualizar un rol existente.",
                "description": "Actualiza un rol existente en la base de datos.",
                "operationId": "1fd69e8ca426ee17dadee1fef3b79e9c",
                "requestBody": {
                    "description": "Datos del rol a actualizar.",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "name",
                                    "permissions"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "Nuevo Nombre de Rol"
                                    },
                                    "permissions": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol actualizado exitosamente."
                    },
                    "422": {
                        "description": "El rol ingresado ya existe o los datos proporcionados son inválidos.",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "El rol ingresado ya existe."
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/deleteRole/{roleId}": {
            "delete": {
                "tags": [
                    "roles"
                ],
                "summary": "Eliminar un rol existente.",
                "description": "Elimina un rol existente en la base de datos.",
                "operationId": "2b30d25dab01012c4680f9d42eae9b0e",
                "responses": {
                    "200": {
                        "description": "Rol eliminado exitosamente."
                    },
                    "404": {
                        "description": "El rol no existe."
                    },
                    "500": {
                        "description": "Error al eliminar el rol."
                    }
                }
            }
        },
...
    }
}